+++ vllm/v1/worker/gpu_model_runner.py
@@ -131,8 +131,10 @@ from vllm.v1.attention.backend import (
     AttentionType,
     CommonAttentionMetadata,
 )
+from vllm.v1.attention.backends.bailing_linear_attn import (
+    BailingLinearAttentionMetadataBuilder,
+)
 from vllm.v1.attention.backends.gdn_attn import GDNAttentionMetadataBuilder
-from vllm.v1.attention.backends.linear_attn import LinearAttentionMetadataBuilder
 from vllm.v1.attention.backends.mamba2_attn import Mamba2AttentionMetadataBuilder
 from vllm.v1.attention.backends.utils import (
     NULL_BLOCK_ID,
@@ -2446,7 +2448,7 @@ class GPUModelRunner(
                 (
                     Mamba2AttentionMetadataBuilder,
                     GDNAttentionMetadataBuilder,
-                    LinearAttentionMetadataBuilder,
+                    BailingLinearAttentionMetadataBuilder,
                 ),
             ):
                 assert ubid is None, (
